refactor(repository): Atualiza tipagem para Integer e adiciona validação de conflito

+++ hospital-management/src/main/java/ufpb/hospital/repository/ResidenteRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ufpb.hospital.model.Residente;
 
-public interface ResidenteRepository extends JpaRepository<Residente, Long> {
+public interface ResidenteRepository extends JpaRepository<Residente, Integer> {
 }
