+++ hospital-management/src/main/java/ufpb/hospital/repository/UnidadeSaudeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ufpb.hospital.model.UnidadeSaude;
 
-public interface UnidadeSaudeRepository extends JpaRepository<UnidadeSaude, Long> {
+public interface UnidadeSaudeRepository extends JpaRepository<UnidadeSaude, Integer> {
 }
